LandonM's Old Website program
Here is the code for the website project I worked on in the past. I don't have access to look at the database, but the username and password are test and 1234 respectively. The database should still work, but is set to reset all the information every so often, so don't be surprised if changes you make to the info are reset.